requirements.txt
timeit==0.3.1
json
requests==2.26.0
numpy
shutil==1.0
os
sys
tkinter==8.6
pandas==1.3.3
matplotlib==3.4.3
pillow==8.3.2
torch==1.9.1
re==2.2.1
base64==1.0
together==1.0
python-dotenv==0.19.1
asyncio==3.4.3
aiohttp==3.7.4